# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |
    API for the WASAText application, which is the project for the WASA course @ Sapienza
  version: 0.0.1
...
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    user_id:
      name: user_id
      description: |
        ID of the user
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 2048
...
  schemas:
    User:
      title: User
      description: |
        Schema defining a user, which can belong to multiple chats and send messages
      type: object
      properties:
        user_id:
          description: ID of the user
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: Maria
        display_name:
          description: Name of the user
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 32
          example: FairyMaria
        biography:
          description: |
            Little biography of the user
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 512
          example: When life gives you lemons...
        profile_pic:
          description: Profile picture
          type: string
          pattern: '^.*$'
          minLength: 0
          maxLength: 44294967296
          format: byte
      required:
        - user_id
        - display_name
        - biography
        - profile_pic
    Message:
      title: Message
      description: |
        Representation of a message
      type: object
      properties:
        message_id:
          description: ID of the message, always expressed in 16 digits
          type: integer
          example: 1
          minimum: 1
          maximum: 1048576 # Could be lower or higher, had to set a maximum
        sender:
          description: ID of the sender
          type: integer
          example: 2
          minimum: 1
          maximum: 2048
        chat:
          description: ID of the chat
          type: integer
          example: 1
          minimum: 1
          maximum: 65536
        content:
          description: Text of the message
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 65536
          example: Lorem ipsum dolor sit est cum magna gratia
        timestamp:
          description: Timestamp of when the message was sent
          type: string
          # format: date-time
          pattern: '^[0-9]*$'
          minLength: 0
          maxLength: 44294967296
          example: '1731321854'
        photo:
          description: Photo sent with the message, if any is sent. Can be null
          type: string
          format: byte
          pattern: '^.*$'
          minLength: 0
          maxLength: 44294967296
        received:
          description: List of users who received the message
          type: array
          minItems: 0
          maxItems: 44294967296
          items:
            $ref: "#/components/schemas/User"
        seen:
          description: List of users who read the message
          type: array
          minItems: 0
          maxItems: 44294967296
          items:
            $ref: "#/components/schemas/User"
        deleted:
          description: Whether the message has been deleted
          type: boolean
        forwarded:
          description: Specifies whether the message was forwarded
          type: boolean
          example: false
        reactions:
          description: Set of reactions applied to the message
          type: array
          minItems: 0
          maxItems: 44294967296
          items:
            $ref: "#/components/schemas/Reaction"
        reply:
          description: ID of the message to which the current replies. 0 means no reply
          type: integer
          example: 1
          minimum: 0
          maximum: 1048576
      required:
        - message_id
        - sender
        - content
        - chat
        - timestamp
        - photo
        - received
        - seen
        - deleted
        - forwarded
        - reactions
        - reply
    Reaction:
      title: Reaction
      description: |
        Defines a reaction, which can be made out of text or an emoji
      type: object
      properties:
        reaction_id:
          description: ID of the reaction
          type: integer
          example: 3
          minimum: 1
          maximum: 1048576
        message:
          description: Message ID to which the reaction corresponds
          type: integer
          example: 2
          minimum: 1
          maximum: 1048576
        sender:
          description: ID of the user
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: Maria
        content:
          description: Content of the reaction
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 16384
      required:
        - reaction_id
        - message
        - sender
        - content
    Chat:
      title: Chat
      description: Representation of a chat
      type: object
      properties:
        chat_id:
          description: ID of the chat
          type: integer
          example: 4
          minimum: 1
          maximum: 65536 # Set to 65536 for convenience, could be higher / lower
        is_private:
          description: Defines whether the chat is private or is a group (if true, the chat is private; if false, the chat is a group)
          type: boolean
          example: true
        users:
          description: List of users participating to the chat
          type: array
          items:
            description: ID of the users
            type: integer
            minimum: 1
            maximum: 2048
          minItems: 0
          maxItems: 4294967296
        name:
          description: Name of the group chat. Visible only if the chat is a group
          type: string
          pattern: '^.*$'
          minLength: 0
          maxLength: 4294967296
          example: The children of Andrew
        group_description:
          description: Description of the group. Visible only if the chat is a group
          type: string
          example: We are such a nice group!
          pattern: '^.*$'
          minLength: 0
          maxLength: 65536
        photo:
          description: Profile picture of the chat. Visible only if the chat is a group
          type: string
          format: byte
          pattern: '^.*$'
          minLength: 0
          maxLength: 4294967296
        messages:
          description: List of all the messages sent in the chat
          type: array
          minItems: 0
          maxItems: 4294967296
          items:
            $ref: "#/components/schemas/Message"
        last_sent:
          description: ID of the last sent message
          type: integer
          example: 2
          minimum: 1
          maximum: 4294967296
      required:
        - chat_id
        - is_private
        - users
        - name
        - group_description
        - photo
        - messages
        - last_sent
security:
  - bearerAuth: []
paths:
  /session:
    # POST = doLogin
    post:
      tags: ["User operations", "Login"]
      summary: Logs in the user
      description: |
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              description: JSON object representing the user
              properties:
                name:
                  type: string
                  description: Name of the user
                  example: Maria
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
        required: true
      
      responses:
        '200':
          description: |
            User log-in action successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        
        '201':
          description: |
            The user does not exist, hence it will be created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        
        '403':
          description: |
            The user is already logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

        '406':
          description: |
            The provided username is too long or too short
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /user/{user_id}/chats:
    # GET = getMyConversations
    parameters:
      - $ref: "#/components/parameters/user_id"

    get:
      tags: ["User operations", "Chats"]
      summary: Get the chats of the user
      description: Get the chats of the user specified in the user ID
      operationId: getMyConversations
      responses:
        "200":
          description: |
            Successfully retrieven the chats of the user
          content:
            application/json:
              schema:
                type: object
                description: Object containing an array of chats to which the user belongs
                properties:
                  user_chats:
                    type: array
                    description: Array containing the chats to which the user belongs
                    items: 
                      $ref: "#/components/schemas/Chat"
                    minItems: 0
                    maxItems: 4294967296
                required: 
                  - user_chats
        
        '404':
          description: |
            The chat does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
  /user/{user_id}/id:
    # POST = setMyUserName
    parameters:
      - $ref: "#/components/parameters/user_id"

    post:
      tags: ["User operations", "Settings"]
      summary: Change the username of the user
      description: Changes the username of the user specified by the ID in the path
      operationId: setMyUserName
      requestBody:
        description: New username
        content:
          application/json:
            schema:
              type: object
              description: Object containing the new username of the user
              properties:
                new_username:
                  type: string
                  description: New ID
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
                  example: Maria
        required: true
      responses:
        "204":
          description: |
            The username has been changed successfully
        "418":
          description: |
            The username wasn't changed due to an error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /user/{user_id}/name:
    # POST = setMyUserName
    parameters:
      - $ref: "#/components/parameters/user_id"

    post:
      tags: [ "User operations", "Settings" ]
      summary: Change the display name of the user
      description: Changes the display name of the user specified by the ID in the path
      operationId: setMyDisplayName
      requestBody:
        description: New display name
        content:
          application/json:
            schema:
              type: object
              description: Object containing the new display name of the user
              properties:
                new_display_name:
                  type: string
                  description: New display name
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 32
                  example: FairyNewMaria
        required: true
      responses:
        "204":
          description: |
            The display name has been changed successfully
        "418":
          description: |
            The display name wasn't changed due to an error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /user/{user_id}/photo:
    # POST = setMyPhoto
    parameters:
      - $ref: "#/components/parameters/user_id"

    post:
      tags: ["User operations"]
      summary: Changes the user's photo
      operationId: setMyPhoto
      description: Changes the user's photo
      requestBody:
        description: The new photo of the user
        content:
          application/json:
            schema:
              type: object
              description: Object containing the new photo of the user
              properties:
                new_photo:
                  type: string
                  description: The new photo of the user
                  minLength: 0
                  maxLength: 4294967296
                  pattern: '^.*$'
                  format: byte
        required: true
      responses:
        "204":
          description: |
            The photo ahs been changed successfully
        
        '403':
          description: |
            You are not the user, you can't change someone else's photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /user/{user_id}/bio:
    # POST = setMyBio
    parameters:
      - $ref: "#/components/parameters/user_id"
    
    post:
      tags: ["User operations", "Settings"]
      summary: Change the bio of the user
      operationId: setMyBio
      description: Change the bio of the user
      requestBody:
        description: The new biography
        content:
          application/json:
            schema:
              description: Object containing the new bio of the user
              type: object
              properties:
                new_bio:
                  type: string
                  description: The new bio of the user
                  pattern: '^.*$'
                  minLength: 1
                  maxLength: 2048
        required: true
      responses:
        "204":
          description: |
            The biography has been changed successfully

        '400':
          description: |
            The biography given does not respect the standards imposed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...